Cleaner internal state.
Previously we tracked the number of started tasks, the number of active tasks,
the number of remaining tasks, and the total number of tasks (tasks.length). Now
we track the number of waiting tasks, the number of active tasks, and the number
of ended tasks. So, started = active + ended, and remaining = waiting + active.

This commit also fixes two minor bugs whereby a task that is deferred within
another task callback can trigger a re-entrant start, and eliminates an
unnecessary error check before deferring a task (since tasks are never started
if there is an error, anyway).