More Promise-like behavior.
If a task synchronously throws an error upon starting, treat this as if the task
were rejected with the error. If a task invokes the callback multiple times,
ignore subsequent invocations rather than throwing an error. In particular, if a
task invokes the callback synchronously, do not allow it to subsequently invoke
the callback asynchronously.

From the Promises A+ specification:

2.3.3.3.3. If both resolvePromise and rejectPromise are called, or multiple
calls to the same argument are made, the first call takes precedence, and any
further calls are ignored.

2.3.3.3.4. If calling then throws an exception e, If resolvePromise or
rejectPromise have been called, ignore it. Otherwise, reject promise with e as
the reason.