Fix #62 - rethrow await callback errors.
This logic is a little bit tricky because the queue should ignore subsequent
errors that are thrown synchronously by a task, but not ignore an error that is
thrown synchronously by the await callback. To identify the former, the queue
now clears the data array before invoking the await callback.